Changes on the style of the web
