+++ turbo-repo/apps/web/app/alpha-2506/[lang]/torre/page.tsx
@@ -2,7 +2,8 @@ import type { Metadata } from "next";
 import Nav from "../../../../components/v2/Nav";
 import ModuleTabs from "../../../../components/v2/ModuleTabs";
 import TorreDeControl from "../../../../components/v2/TorreDeControl";
-import { Footer, FinalCta } from "../../../../components/v2/Sections";
+import { Footer } from "../../../../components/v2/sections/Footer";
+import { FinalCta } from "../../../../components/v2/sections/FinalCta";
 
 export const metadata: Metadata = {
   title: "Torre de control — ModularIoT",
